% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcovU.test.R
\name{dcovU.test}
\alias{dcovU.test}
\title{Permutation bootstrap methods for independence test of unbiased distance covariance.}
\usage{
dcovU.test(X, Y, R = 1000, alpha = 0.05)
}
\arguments{
\item{X}{\eqn{N-by-p} matrix. The sample dataset of X. Each row represents one observation.}

\item{Y}{\eqn{N-by-q} matrix. The sample dataset of Y. Each row represents one observation.}

\item{R}{Number of replicates for independence test. Default is set to 1000.}

\item{alpha}{The confidence level. The default value is 0.05.}
}
\value{
A list containing the following components will be returned:
\describe{
\item{threshold}{The threshold dcovU value of given confidence interval;}
\item{estimate}{The estimated unbiased distance covariance square;}
}
}
\description{
\code{dcovU.test} is used to do perform independence test via permutation bootstrap methods, which perform permutations for empirical distribution function.
The distance measure is set as Euclideam measure.
}
\examples{
library(MASS)
Num = 40
p = 3
q = 2
mu1 = rep(0,p)
sigma1 = diag(p)
mu2 = rep(0,q)
sigma2 = diag(q)
X = mvrnorm(Num,mu1,sigma1)
Y = mvrnorm(Num,mu2,sigma2)
SUM_X = rowSums(X)
Y[,1] = SUM_X+0.01*Y[,1]
Y[,2] = (SUM_X)^2+0.01*Y[,2]
\donttest{
dcovU.test(X,Y)
}
}
