% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mminus.R
\name{Mminus}
\alias{Mminus}
\title{Calculate the M minus matrix.}
\usage{
Mminus(gM, Xdiff)
}
\arguments{
\item{gM}{\eqn{N(N-1)/2-by-q} matrix. The g-Matrix of Y.
It can be obtained by function \code{gMatrix}.}

\item{Xdiff}{\eqn{N(N-1)/2-by-p} matrix. The matrix of the pairwise difference of X.
Each row represents one observation. It can be obtained by function \code{Xdiff}.}
}
\value{
The M minus matrix.
}
\description{
\code{Mminus} is used to calculate the M minus matrix defined in our paper.
}
\examples{
library(MASS)
Num = 40
p = 3
q = 2
mu1 = rep(0,p)
sigma1 = diag(p)
mu2 = rep(0,q)
sigma2 = diag(q)
X = mvrnorm(Num,mu1,sigma1)
Y = mvrnorm(Num,mu2,sigma2)
SUM_X = rowSums(X)
Y[,1] = SUM_X+0.01*Y[,1]
Y[,2] = (SUM_X)^2+0.01*Y[,2]
X_diff <- Xdiff(X)
gM_Y <- gMatrix(Y)
g_Y <- gM_Y$gM
Mminus_matrix <- Mminus(g_Y,X_diff)
}
