% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2D2R.R
\name{S2D2R}
\alias{S2D2R}
\title{Get the dimension reduction subspaces for both X and Y.}
\usage{
S2D2R(
  X,
  Y,
  e = 10^-3,
  xi = 200,
  psi = 10,
  rho = 1000,
  omega = 10,
  abs.epsilon = 10^-3,
  rel.epsilon = 10^-3,
  alpha = 0.05,
  R = 1000
)
}
\arguments{
\item{X}{\eqn{N-by-p} matrix. The sample dataset of X. Each row represents one observation.}

\item{Y}{\eqn{N-by-q} matrix. The sample dataset of Y. Each row represents one observation.}

\item{e}{The tolerance of two values calculated by adjacent two steps by Difference of Convex Algorithm (DCA).
If the two values is smaller than e, we terminate the loop of DCA.
The default value is 10^-3.}

\item{xi}{One of the augmented Lagrangian parameters.
It should be nonnegative as well as bigger than psi.
The default value is 200.}

\item{psi}{The other augmented Lagrangian parameters.
The default value is 10.}

\item{rho}{The augmented Lagrangian parameter in subproblem.
The default value is 1000.}

\item{omega}{The increment of xi. It should be bigger than 1.
The default value is 10.}

\item{abs.epsilon}{The absolute tolerance in ADMM.
The default value is 10^-3.}

\item{rel.epsilon}{The relative tolerance in ADMM.
The default value is 10^-3.}

\item{alpha}{The confidence level. The default value is 0.05.}

\item{R}{Number of replicates in permutation bootstrap method for independence test.}
}
\value{
A list containing the following components will be returned:
\describe{
\item{dW1}{The dimension of subspace W1;}
\item{basis.W1}{The basis of subspace W1;}
\item{dW2}{The dimension of subspace W2;}
\item{basis.W2}{The basis of subspace W2}
}
}
\description{
\code{S2D2R} is used to do dimension reduction via distance covariance by projecting X and Y into subspaces \eqn{W_1,} and \eqn{W_2}, respectively.
The dimension of X and Y, \eqn{W_1,} and \eqn{W_2} can be different.
}
\examples{
library(MASS)
Num = 40
p = 3
q = 2
mu1 = rep(0,p)
sigma1 = diag(p)
mu2 = rep(0,q)
sigma2 = diag(q)
X = mvrnorm(Num,mu1,sigma1)
Y = mvrnorm(Num,mu2,sigma2)
SUM_X = rowSums(X)
Y[,1] = SUM_X+0.01*Y[,1]
Y[,2] = (SUM_X)^2+0.01*Y[,2]
\donttest{
S2D2R(X,Y)
}
}
